Reorganized Functions Python Reorganized Functions in python backend to be grouped in files based on functionality rather all in one main file.
