+++ adversarial_apps/src/app/api/call-python-api/python-calls/main_api.py
@@ -9,709 +9,18 @@ import requests
 from bs4 import BeautifulSoup
 from dotenv import load_dotenv
 
-
-def sanitize_search_term(search_term: str) -> str:
-
-    # Validate the search term (e.g., allow only alphanumeric and a few specific characters)
-    if not re.match(r"^[A-Za-z0-9&\-.,\s]+$", search_term):
-        raise ValueError(
-            "Invalid search term: only alphanumeric and a few special characters allowed."
-        )
-
-    # Sanitize the input by removing unsafe characters, Allow only alphanumeric, &, -, ., ,, and spaces
-    sanitized_term = re.sub(r"[^A-Za-z0-9&\-., ]", "", search_term)
-
-    # Normalize multiple spaces to a single space
-    sanitized_term = re.sub(r"\s+", " ", sanitized_term)
-
-    # escape for HTML contexts to prevent XSS
-    sanitized_term = html.escape(sanitized_term)
-
-    return sanitized_term
-
-
-def sanitize_and_validate_cik(cik: str) -> str:
-    """
-    Validates and sanitizes the CIK input.
-
-    :param cik: Input string representing the CIK number.
-    :return: Sanitized CIK number if valid, raises ValueError otherwise.
-    """
-    # Trim leading and trailing whitespace
-    cik = cik.strip()
-
-    # Validate the CIK format: must be only digits
-    if not re.match(r"^\d+$", cik):
-        raise ValueError("Invalid CIK: must be only digits.")
-
-    # Return the sanitized and validated CIK
-    return cik
-
-
-def obtain_cik_number(search_term: str) -> dict:
-    """
-    Returns list of companies based on search and their CIK numbers
-    :param search_term: string to search for
-    :return: dictionary with list of companies and their CIK numbers
-    """
-    try:
-        # Set up the payload to make the search request
-        url = "https://www.sec.gov/cgi-bin/cik_lookup"
-        headers = {
-            "Content-Type": "application/x-www-form-urlencoded",
-            "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8",
-            "User-Agent": "JamesAllen <ja799793@ucf.edu> (Adversarial Apps)",
-            "Referer": "https://www.sec.gov/search-filings/cik-lookup",
-        }
-
-        # The term to search for
-        payload = {"company": sanitize_search_term(search_term)}
-
-        response = requests.post(url, headers=headers, data=payload)
-
-        # If the request was successful (status code 200)
-        if response.status_code == 200:
-            # Parse the HTML to obtain the CIK number and company name
-            soup = BeautifulSoup(response.text, "html.parser")
-
-            # Find the <pre> tag that contains the CIK and company data
-            pre_tag = soup.find_all("pre")
-
-            cik_data = []
-
-            # Iterate over each <pre> tag and extract CIK and company name
-            for pre in pre_tag:
-                # Regular expression to match CIK and company name
-                matches = re.findall(
-                    r"(\d{10})\s+(.*?)\s*$", pre.get_text(), re.MULTILINE
-                )
-                for cik, company in matches:
-                    cik_data.append({"CIK": cik, "Company Name": company})
-
-            # Return the result as a dictionary
-            return {"status": "success", "companies": cik_data}
-        else:
-            return {
-                "status": "error",
-                "message": f"Failed to fetch data: {response.status_code} {response.text}",
-            }
-    except ValueError as e:
-        return {"status": "error", "message": f"Invalid input: {str(e)}"}
-    except Exception as e:
-        return {"status": "error", "message": f"An error occurred: {str(e)}"}
-
-
-def get_sec_data(cik: str) -> dict:
-    """
-    Retrieve SEC json file for company based on given CIK number
-
-    :param cik: CIK number for company to retrieve data for
-    :return: dictionary with json data
-    """
-
-    try:
-        sanitized_cik = sanitize_and_validate_cik(cik)
-    except ValueError as e:
-        return {"status": "error", "message": f"{str(e)}"}
-
-    url = f"https://data.sec.gov/submissions/CIK{sanitized_cik}.json"
-
-    headers = {
-        "User-Agent": "JamesAllen <ja799793@ucf.edu> (Adversarial Apps)",
-        "Accept-Encoding": "gzip, deflate",
-        "Host": "data.sec.gov",
-    }
-    response = requests.get(url, headers=headers)
-
-    if response.status_code == 200:
-        try:
-            # Parse the JSON response
-            data = response.json()
-
-            # Extract the required fields
-            company_data = {
-                "name": data.get("name", "N/A"),  # Company name
-                "formerNames": data.get("formerNames", []),  # Former names array
-                "address": data.get("addresses", {})
-                .get("business", {})
-                .get("street1", "N/A"),
-                "street2": data.get("addresses", {})
-                .get("business", {})
-                .get("street2", "N/A"),
-                "city": data.get("addresses", {})
-                .get("business", {})
-                .get("city", "N/A"),
-                "zipCode": data.get("addresses", {})
-                .get("business", {})
-                .get("zipCode", "N/A"),
-                "stateOrCountryDescription": data.get("addresses", {})
-                .get("business", {})
-                .get("stateOrCountryDescription", "N/A"),
-                "stateOfIncorporation": data.get(
-                    "stateOfIncorporationDescription", "N/A"
-                ),
-                "phone": data.get("phone", "N/A"),
-                "website": data.get("website", "N/A"),  # Assume there's a website field
-            }
-
-            # Get all filing dates from the filings section
-            filing_dates = (
-                data.get("filings", {}).get("recent", {}).get("filingDate", [])
-            )
-
-            # If there are no filing dates, return 'N/A'
-            if not filing_dates:
-                recent_filing_date = ""
-            else:
-                # Sort the filing dates in descending order to get the most recent one
-                filing_dates_sorted = sorted(filing_dates, reverse=True)
-                recent_filing_date = filing_dates_sorted[0]
-
-            # Add the most recent filing date to the company data
-            company_data["mostRecentFilingDate"] = recent_filing_date
-
-            # Ensure proper formatting of `formerNames`
-            if "formerNames" in company_data and company_data["formerNames"]:
-
-                company_data["formerNames"] = [
-                    {
-                        "name": former.get("name", "N/A"),
-                        "fromDate": former.get("from", "N/A"),
-                        "toDate": former.get("to", "N/A"),
-                    }
-                    for former in company_data["formerNames"]
-                ]
-
-            return {"status": "success", "company": company_data}
-
-        except KeyError as e:
-            return {"status": "error", "message": f"KeyError: {str(e)}"}
-        except Exception as e:
-            return {"status": "error", "message": f"An error occurred: {str(e)}"}
-    else:
-        return {
-            "status": "error",
-            "message": f"Unable to retrieve data for CIK {sanitized_cik} (Status Code: {response.status_code})",
-        }
-
-
-def add_user(username: str, password_hashed: str, company: str) -> dict:
-    """
-    Add user to the system
-
-    :param username: username of user
-    :param password_hashed: hashed password of user
-    :param company: company of user
-    :return: message indicating success or failure
-    """
-
-    connection = None
-
-    try:
-        # Connect to the PostgreSQL database using the URI
-        connection = psycopg2.connect(os.getenv("DATABASE_URL"))
-        cursor = connection.cursor()
-
-        # Check if the username exists
-        cursor.execute('SELECT username FROM "USERS" WHERE username = %s', (username,))
-        username_exists = cursor.fetchone()  # fetches the first result from query
-
-        # if we get a result from the query, means we have a at least 1 user already with username
-        if username_exists:
-            return {
-                "status": "error",
-                "message": f"Username '{username}' already exists.",
-            }
-        else:
-            # if nothing, then we can add user since still unique
-
-            # Query, which inserts user info with information provided
-            # if username is in database, we get no result
-            query = """
-            INSERT INTO "public" . "USERS" (username, password, company, "isReviewer")
-            VALUES (%s, %s, %s, %s)
-            ON CONFLICT (username) DO NOTHING
-            RETURNING username;
-            """
-
-            # Execute the query with the provided data
-            cursor.execute(query, (username, password_hashed, company, False))
-
-            # RETURNING in query returns username, this gets the username from that return
-            result = cursor.fetchone()
-
-            if result:
-                # Finalizes the query to the database to be saved
-                connection.commit()
-                return {
-                    "status": "success",
-                    "message": f"User {result[0]} added successfully",
-                }
-            else:
-                return {
-                    "status": "error",
-                    "message": f"No user inserted (conflict detected).",
-                }
-
-    except psycopg2.Error as e:
-        print(f"Database error: {e}")
-    finally:
-        # Ensure the connection is closed, as finished with code
-        if connection:
-            cursor.close()
-            connection.close()
-
-
-def get_password(username: str) -> dict:
-    """
-    Retrieve hashed password for user based on username
-
-    :param username: username of user
-    :return: dictionary with hashed password
-    """
-
-    connection = None
-
-    try:
-        # Connect to the PostgreSQL database using the URI
-        connection = psycopg2.connect(os.getenv("DATABASE_URL"))
-        cursor = connection.cursor()
-
-        # Query to get the hashed password for the provided username
-        cursor.execute('SELECT password FROM "USERS" WHERE username = %s', (username,))
-        password = cursor.fetchone()
-
-        if password:
-            return {"status": "success", "password": password[0]}
-        else:
-            return {"status": "error", "message": f"Username '{username}' not found."}
-
-    except psycopg2.Error as e:
-        print(f"Database error: {e}")
-    finally:
-        # Ensure the connection is closed, as finished with code
-        if connection:
-            cursor.close()
-            connection.close()
-
-    return {
-        "status": "error",
-        "message": f"An error occurred retrieving password for '{username}'",
-    }
-
-
-def get_reviewer_status(username: str) -> dict:
-    """
-    Retrieve reviewer status for user based on username
-
-    :param username: username of user
-    :return: dictionary with reviewer status value
-    """
-    connection = None
-
-    try:
-        # Connect to the PostgreSQL database using the URI
-        db_url = os.getenv("DATABASE_URL")
-
-        connection = psycopg2.connect(db_url)
-        cursor = connection.cursor()
-
-        # Query to get the hashed password for the provided username
-        cursor.execute(
-            'SELECT "isReviewer" FROM "USERS" WHERE username = %s', (username,)
-        )
-        reviewerStatus = cursor.fetchone()
-
-        if reviewerStatus:
-            return {"status": "success", "reviewerStatus": reviewerStatus[0]}
-        else:
-            return {"status": "error", "message": f"Username '{username}' not found."}
-
-    except psycopg2.Error as e:
-        return {"status": "error", "message": str(e)}
-    finally:
-        # Ensure the connection is closed, as finished with code
-        if connection:
-            cursor.close()
-            connection.close()
-
-    return {
-        "status": "error",
-        "message": f"An error occurred retrieving reviewer status for '{username}'",
-    }
-
-
-def add_remove_favorite(username: str, cik: int) -> dict:
-    """
-    Add or remove a favorite company for the user.
-
-    :param username: Username of the user
-    :param cik: CIK number of the company
-    :return: Message indicating success or failure
-    """
-    try:
-        with psycopg2.connect(os.getenv("DATABASE_URL")) as connection:
-            with connection.cursor() as cursor:
-
-                user_id = get_user_id(username, cursor)
-
-                if not user_id:
-                    return {
-                        "status": "error",
-                        "message": f"User '{username}' not found.",
-                    }
-
-                # company must exist in the database first before we can add it to favorites
-                cursor.execute('SELECT 1 FROM "COMPANIES" WHERE "CIK" = %s', (cik,))
-                company_exists = cursor.fetchone()
-
-                if not company_exists:
-                    # since company does not exist, we need to add it to the database before we can add it to favorites
-                    # do the the CIK in FAVORITES table is a foreign key to the CIK in COMPANIES table
-                    cursor.execute(
-                        'INSERT INTO "COMPANIES" ("CIK", "isVerified", "riskScore") VALUES (%s, %s, %s)',
-                        (cik, False, 0),
-                    )
-                    connection.commit()
-
-                # Check if the user has already favorited the company
-                cursor.execute(
-                    'SELECT 1 FROM "FAVORITES" WHERE "userId" = %s AND "companyCIK" = %s',
-                    (user_id, cik),
-                )
-                favorite_exists = cursor.fetchone()
-
-                if favorite_exists:
-                    # If the favorite exists, remove it
-                    cursor.execute(
-                        'DELETE FROM "FAVORITES" WHERE "userId" = %s AND "companyCIK" = %s',
-                        (user_id, cik),
-                    )
-                    connection.commit()
-                    return {
-                        "status": "success",
-                        "message": f"Removed company with CIK {cik} from favorites for {username}.",
-                    }
-                else:
-                    # If the favorite does not exist, add it
-                    cursor.execute(
-                        'INSERT INTO "FAVORITES" ("userId", "companyCIK") VALUES (%s, %s)',
-                        (user_id, cik),
-                    )
-                    connection.commit()
-                    return {
-                        "status": "success",
-                        "message": f"Added company with CIK {cik} to favorites for {username}.",
-                    }
-    except psycopg2.Error as e:
-        return {"status": "error", "message": f"Database error: {e}"}
-
-
-def get_favorites(username: str) -> dict:
-    """
-    Retrieve the list of favorited companies for the user.
-
-    :param username: Username of the user
-    :return: Dictionary containing the list of favorited companies
-    """
-    try:
-        with psycopg2.connect(os.getenv("DATABASE_URL")) as connection:
-            with connection.cursor() as cursor:
-
-                user_id = get_user_id(username, cursor)
-
-                # If the user does not exist, return an error message
-                if not user_id:
-                    return {
-                        "status": "error",
-                        "message": f"User '{username}' not found.",
-                    }
-
-                # Query to get the list of favorited companies
-                cursor.execute(
-                    'SELECT "companyCIK" FROM "FAVORITES" WHERE "userId" = %s',
-                    (user_id,),
-                )
-                favorites = cursor.fetchall()
-
-                return {
-                    "status": "success",
-                    "favorites": [favorite[0] for favorite in favorites],
-                }
-    except psycopg2.Error as e:
-        return {"status": "error", "message": f"Database error: {e}"}
-
-
-def get_user_id(username: str, cursor) -> int:
-    """
-    Retrieve the user ID for the provided username.
-
-    :param username: Username of the user
-    :param cursor: Cursor object to execute the query (already connected to the database)
-    :return: User ID
-    """
-
-    try:
-        cursor.execute('SELECT id FROM "USERS" WHERE username = %s', (username,))
-        user = cursor.fetchone()
-
-        # if the entry is not found, then no user so no ID
-        if not user:
-            return None
-        # the id is the first element in the user row, which is what was fetched
-        return user[0]
-    except psycopg2.Error as e:
-        return None
-
-
-def get_company_score(cik: str) -> dict:
-    """
-    Retrieve the risk score for a verified company based on the CIK number.
-
-    :param cik: CIK number of the company
-    :return: Dictionary containing the risk score
-    """
-    try:
-        with psycopg2.connect(os.getenv("DATABASE_URL")) as connection:
-            with connection.cursor() as cursor:
-
-                # Combined query to check if the company exists, is verified, and retrieve riskScore
-                cursor.execute(
-                    'SELECT "isVerified", "riskScore" FROM "COMPANIES" WHERE "CIK" = %s',
-                    (cik,),
-                )
-                result = cursor.fetchone()
-
-                # If the company is not found
-                if not result:
-                    return {
-                        "status": "error",
-                        "message": f"Company with CIK {cik} not found.",
-                    }
-
-                is_verified, risk_score = result
-
-                # the score must be verified and reviewed before being returned
-                if not is_verified:
-                    return {
-                        "status": "error",
-                        "message": f"Company with CIK {cik} is not verified.",
-                    }
-
-                # If company is verified, return the risk score
-                return {
-                    "status": "success",
-                    "riskScore": risk_score,
-                }
-
-    except psycopg2.Error as e:
-        return {"status": "error", "message": f"Database error: {e}"}
-
-
-def get_recent_ownerships(cik: str, pagination: int) -> dict:
-    """
-    Retrieve recent ownerships for a company based on the CIK number.
-
-    :param cik: CIK number of the company
-    :param pagination: Index for pagination
-    :return: Dictionary containing the list of recent ownerships
-    """
-
-    # obtains the .json file from the SEC website
-    url = f"https://data.sec.gov/submissions/CIK{cik}.json"
-
-    headers = {
-        "User-Agent": "JamesAllen <ja799793@ucf.edu> (Adversarial Apps)",
-        "Accept-Encoding": "gzip, deflate",
-        "Host": "data.sec.gov",
-    }
-
-    response = requests.get(url, headers=headers)
-
-    if response.status_code == 200:
-        try:
-            # Parse the JSON response to grab form information
-            data = response.json()
-
-            recent = data.get("filings", {}).get("recent", {})
-            forms = recent.get("form", [])
-            filling_dates = recent.get("filingDate", [])
-            accession_numbers = recent.get("accessionNumber", [])
-
-            # Grabs all form 4 forms
-            form_indices = [i for i, form in enumerate(forms) if form == "4"]
-
-            # Connects forms 4 with their respective filling dates and accession numbers
-            # for ease of reference
-            filtered_forms = [
-                {
-                    "fillingDate": filling_dates[i],
-                    "form": forms[i],
-                    "accessionNumber": accession_numbers[i],
-                }
-                for i in form_indices
-            ]
-
-            # Paginate: Get the first 5 items for the given page
-            start_index = (pagination - 1) * 5
-            end_index = start_index + 5
-
-            # for each of the forms within the index range, grab data file and process xml
-            for form in filtered_forms[start_index:end_index]:
-                # each form stored at unique url based off cik without zeros
-                # and modified accession number in url following pattern below
-                url = f"https://www.sec.gov/Archives/edgar/data/{int(cik)}/{form['accessionNumber'].replace('-', '')}/{form['accessionNumber']}.txt"
-
-                headers = {
-                    "User-Agent": "JamesAllen <ja799793@ucf.edu> (Adversarial Apps)",
-                    "Accept-Encoding": "gzip, deflate",
-                    "Host": "www.sec.gov",
-                }
-
-                response = requests.get(url, headers=headers)
-
-                # if the request was successful, then we can parse the xml
-                if response.status_code == 200:
-                    # grabs all the content within <XML> tag which form 4 is stored in
-                    xml_content = re.search(
-                        r"<XML>(.*?)</XML>", response.text, re.DOTALL
-                    )
-
-                    # if the xml content is found, then we can parse it, and store it in the form dictionary to be returned
-                    if xml_content:
-
-                        # used to get the xml from the extraction above
-                        xml_string = xml_content.group(1)
-
-                        # uses beatifulsoup to parse the xml content
-                        soup = BeautifulSoup(xml_string, "lxml-xml")
-
-                        # grabs the owner information from the xml
-                        issuer = soup.find("issuerName").text
-                        reporter = soup.find("rptOwnerName").text
-
-                        # adds gathered information to the form dictionary to be saved to output
-                        form["issuer"] = issuer
-                        form["reporter"] = reporter
-
-                else:
-                    return {
-                        "status": "error",
-                        "message": f"Unable to retrieve ownership data for CIK {cik} (Status Code: {response.status_code})",
-                    }
-
-            # Return the paginated list of recent ownerships
-            return {
-                "status": "success",
-                "recentOwnerships": filtered_forms[start_index:end_index],
-            }
-
-        except Exception as e:
-            return {"status": "error", "message": f"An error occurred: {str(e)}"}
-    else:
-        return {
-            "status": "error",
-            "message": f"Unable to retrieve recent fillings for CIK {cik} (Status Code: {response.status_code})",
-        }
-
-
-def verify_company(cik: str) -> dict:
-    """
-    Verify a company by its CIK. If the CIK doesn't exist in COMPANIES,
-    insert a new row with isVerified = TRUE. If it does exist and isVerified
-    is FALSE, update it to TRUE.
-
-    :param cik: The CIK number of the company
-    :return: A dictionary with status and message
-    """
-    connection = None
-    try:
-        connection = psycopg2.connect(os.getenv("DATABASE_URL"))
-        cursor = connection.cursor()
-
-        # 1. Check if the company exists in COMPANIES
-        cursor.execute(
-            'SELECT "CIK", "isVerified" FROM "COMPANIES" WHERE "CIK" = %s', (cik,)
-        )
-        row = cursor.fetchone()
-
-        if row:
-            # row = (CIK, isVerified)
-            _, is_verified = row
-            if not is_verified:
-                # 2a. If company exists but isVerified is FALSE, set it to TRUE
-                cursor.execute(
-                    'UPDATE "COMPANIES" SET "isVerified" = TRUE WHERE "CIK" = %s',
-                    (cik,),
-                )
-                connection.commit()
-                return {
-                    "status": "success",
-                    "message": f"Company with CIK {cik} has been verified (updated).",
-                }
-            else:
-                # If it's already verified, just inform the caller
-                return {
-                    "status": "success",
-                    "message": f"Company with CIK {cik} is already verified.",
-                }
-        else:
-            # 2b. If the company doesn't exist, insert it with isVerified = TRUE
-            cursor.execute(
-                'INSERT INTO "COMPANIES" ("CIK", "isVerified", "riskScore") VALUES (%s, %s, %s)',
-                (cik, True, 0),
-            )
-            connection.commit()
-            return {
-                "status": "success",
-                "message": f"Company with CIK {cik} added and verified.",
-            }
-
-    except psycopg2.Error as e:
-        print(f"Database error: {e}")
-        return {"status": "error", "message": f"Database error: {e}"}
-    finally:
-        if connection:
-            cursor.close()
-            connection.close()
-
-
-def update_company_score (cik: str, risk_score: int) -> dict:
-    """
-    Update the risk score for a company based on the CIK number.
-
-    :param cik: CIK number of the company
-    :param risk_score: New risk score for the company
-    :return: Dictionary containing the updated risk score
-    """
-    try:
-        with psycopg2.connect(os.getenv("DATABASE_URL")) as connection:
-            with connection.cursor() as cursor:
-
-                # checks if company is in the database and if not adds them
-                cursor.execute('SELECT 1 FROM "COMPANIES" WHERE "CIK" = %s', (cik,))
-                company_exists = cursor.fetchone()
-
-                if not company_exists:
-                    cursor.execute(
-                        'INSERT INTO "COMPANIES" ("CIK", "isVerified", "riskScore") VALUES (%s, %s, %s)',
-                        (cik, False, 0),
-                    )
-                    connection.commit()
-
-                # Update the risk score for the company and sets it to be verified
-                cursor.execute(
-                    'UPDATE "COMPANIES" SET "riskScore" = %s, "isVerified" = TRUE WHERE "CIK" = %s',
-                    (risk_score, cik),
-                )
-
-                return {
-                    "status": "success",
-                    "message": f"Risk score updated for company with CIK {cik}.",
-                }
-
-    except psycopg2.Error as e:
-        return {"status": "error", "message": f"Database error: {e}"}
+from search import obtain_cik_number
+from edgar import get_sec_data, get_recent_ownerships
+from database import (
+    add_user,
+    get_password,
+    get_reviewer_status,
+    add_remove_favorite,
+    get_favorites,
+    get_company_score,
+    verify_company,
+    update_company_score,
+)
 
 # the call-python-api will call it here, and provides the inputActionAndData
 # which then determines which part of the API to run
